+++ src/Firebase/Auth/UserRecord.php
@@ -5,13 +5,9 @@ declare(strict_types=1);
 namespace Kreait\Firebase\Auth;
 
 use DateTimeImmutable;
-use Kreait\Firebase\Util\Deprecation;
 use Kreait\Firebase\Util\DT;
 use Kreait\Firebase\Util\JSON;
 
-/**
- * @property array<string, mixed> $customAttributes Deprecated, use {@see UserRecord::$customClaims} instead
- */
 class UserRecord implements \JsonSerializable
 {
     public string $uid;
@@ -107,12 +103,6 @@ class UserRecord implements \JsonSerializable
      */
     public function __get(string $name)
     {
-        if (\mb_strtolower($name) === 'customattributes') {
-            Deprecation::trigger(__CLASS__.'::customAttributes', __CLASS__.'::customClaims');
-
-            return $this->customClaims;
-        }
-
         return $this->{$name};
     }
 }