+++ src/Firebase/Http/Requests.php
@@ -21,16 +21,6 @@ final class Requests implements IteratorAggregate
         $this->requests = $requests;
     }
 
-    /**
-     * @deprecated 5.14.0
-     */
-    public function findBy(callable $callable): ?RequestInterface
-    {
-        $results = \array_filter($this->requests, $callable);
-
-        return \array_shift($results) ?: null;
-    }
-
     public function findByContentId(string $contentId): ?RequestInterface
     {
         foreach ($this->requests as $request) {